+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Simplified status variable check
 - Simplified context merge
 - Fixed invalid looking at wrong variable
-- Renamed `fault` and `thrower` to `origin` and `source` respectively
+- Renamed `fault` and `thrower` to `caused_by` and `thrown_by` respectively
 - Removed unused `additional_result_data` method
 ### Removed
 - Removed context init