+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.3] - 2024-09-30
+### Changed
+- Simplify error building
+- Reduced recalling error since we can just throw it once
+
 ## [2.0.2] - 2024-09-29
 ### Added
 - faultable module